Move MAPI error message formatter to a common location
