+++ src/components/MAPI/clusters/ClusterDetail/ClusterDetailWidgetLabels.tsx
@@ -3,7 +3,7 @@ import ClusterLabels from 'Cluster/ClusterDetail/ClusterLabels/ClusterLabels';
 import ErrorReporter from 'lib/errors/ErrorReporter';
 import { FlashMessage, messageTTL, messageType } from 'lib/flashMessage';
 import { useHttpClientFactory } from 'lib/hooks/useHttpClientFactory';
-import { extractErrorMessage } from 'MAPI/organizations/utils';
+import { extractErrorMessage } from 'MAPI/utils';
 import * as capiv1alpha3 from 'model/services/mapi/capiv1alpha3';
 import PropTypes from 'prop-types';
 import React, { useMemo, useState } from 'react';
