+++ src/components/MAPI/clusters/ClusterList/ClusterListItem.tsx
@@ -7,7 +7,7 @@ import ErrorReporter from 'lib/errors/ErrorReporter';
 import { relativeDate } from 'lib/helpers';
 import { useHttpClientFactory } from 'lib/hooks/useHttpClientFactory';
 import RoutePath from 'lib/routePath';
-import { extractErrorMessage } from 'MAPI/organizations/utils';
+import { extractErrorMessage } from 'MAPI/utils';
 import {
   fetchNodePoolListForCluster,
   fetchNodePoolListForClusterKey,
