+++ src/components/MAPI/clusters/CreateCluster/index.tsx
@@ -6,8 +6,8 @@ import ErrorReporter from 'lib/errors/ErrorReporter';
 import { FlashMessage, messageTTL, messageType } from 'lib/flashMessage';
 import { useHttpClientFactory } from 'lib/hooks/useHttpClientFactory';
 import RoutePath from 'lib/routePath';
-import { extractErrorMessage } from 'MAPI/organizations/utils';
 import { Cluster, ControlPlaneNode, ProviderCluster } from 'MAPI/types';
+import { extractErrorMessage } from 'MAPI/utils';
 import { generateUID } from 'MAPI/utils';
 import * as releasev1alpha1 from 'model/services/mapi/releasev1alpha1';
 import React, { useEffect, useReducer, useRef } from 'react';
