+++ src/components/MAPI/organizations/utils.ts
@@ -1,7 +1,5 @@
 import { Cluster } from 'MAPI/types';
-import { GenericResponse } from 'model/clients/GenericResponse';
 import * as capiv1alpha3 from 'model/services/mapi/capiv1alpha3';
-import * as metav1 from 'model/services/mapi/metav1';
 
 /**
  * Compute an organization namespace from the given organization name.
@@ -9,32 +7,6 @@ import * as metav1 from 'model/services/mapi/metav1';
  */
 export function getOrgNamespaceFromOrgName(name: string) {
   return `org-${name}` as const;
-}
-
-/**
- * Extract the error message from a K8s API response.
- * @param fromErr
- * @param fallback - What message to return if the message
- * could not be extracted.
- */
-export function extractErrorMessage(
-  fromErr: unknown,
-  fallback = 'Something went wrong'
-): string {
-  if (!fromErr) return '';
-
-  let message = '';
-
-  if (metav1.isStatus((fromErr as GenericResponse).data)) {
-    message =
-      (fromErr as GenericResponse<metav1.IK8sStatus>).data.message ?? '';
-  } else if (fromErr instanceof Error) {
-    message = fromErr.message;
-  }
-
-  message ||= fallback;
-
-  return message;
 }
 
 export enum OrganizationNameStatusMessage {
