+++ src/components/MAPI/workernodes/WorkerNodesCreateNodePool.tsx
@@ -4,7 +4,6 @@ import produce from 'immer';
 import ErrorReporter from 'lib/errors/ErrorReporter';
 import { FlashMessage, messageTTL, messageType } from 'lib/flashMessage';
 import { useHttpClientFactory } from 'lib/hooks/useHttpClientFactory';
-import { extractErrorMessage } from 'MAPI/organizations/utils';
 import {
   BootstrapConfig,
   Cluster,
@@ -12,6 +11,7 @@ import {
   ProviderCluster,
   ProviderNodePool,
 } from 'MAPI/types';
+import { extractErrorMessage } from 'MAPI/utils';
 import {
   generateUID,
   getClusterReleaseVersion,
