+++ src/components/MAPI/workernodes/WorkerNodesNodePoolItemScale.tsx
@@ -3,8 +3,8 @@ import { Box, Text } from 'grommet';
 import ErrorReporter from 'lib/errors/ErrorReporter';
 import { FlashMessage, messageTTL, messageType } from 'lib/flashMessage';
 import { useHttpClientFactory } from 'lib/hooks/useHttpClientFactory';
-import { extractErrorMessage } from 'MAPI/organizations/utils';
 import { NodePool } from 'MAPI/types';
+import { extractErrorMessage } from 'MAPI/utils';
 import { getNodePoolScaling } from 'MAPI/utils';
 import PropTypes from 'prop-types';
 import React, { useMemo, useState } from 'react';
